[Refactor] 설정 버튼 트리거인 setter 함수 드래그 영역으로 삽입
