+++ src/pages/mydashboard.tsx
@@ -79,7 +79,7 @@ export default function MyDashboard() {
         message={message}
       />
       <div className="flex flex-col py-6 px-6 tablet:py-10 tablet:px-10 gap-6 tablet:gap-12 laptop:gap-10 max-w-[1022px]">
-        <EditCard
+        {/* <EditCard
           setIsCardEdit={setIsCardEdit}
           isCardEdit={isCardEdit}
           cardId={11808}
@@ -89,9 +89,9 @@ export default function MyDashboard() {
           cardTitle="부산으로 야반도주"
           ModalOpenButton="테스트"
           setIsCardEdit={setIsCardEdit}
-        >
-          <CardModal cardId={11808} columnId={46358} columnTitle="To Do" />
-        </DetailContent>
+        > */}
+        {/* <CardModal cardId={11808} columnId={46358} columnTitle="To Do" /> */}
+        {/* </DetailContent> */}
         <CreateCard columnId={46360}></CreateCard>
         <div className="flex flex-col gap-4 laptop:gap-3">
           <div className="w-full grid gap-[10px] grid-cols-1 tablet:grid-cols-2  laptop:grid-cols-3">
